[JENKINS-38313] - Squash LogStorage code into a single class
