--- core/src/main/java/jenkins/model/logging/impl/FileLogCompatLayer.java
@@ -1,55 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2018 CloudBees, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package jenkins.model.logging.impl;
-
-import hudson.AbortException;
-import jenkins.model.logging.Loggable;
-import org.kohsuke.accmod.Restricted;
-import org.kohsuke.accmod.restrictions.Beta;
-
-import javax.annotation.Nonnull;
-import java.io.File;
-import java.io.IOException;
-
-/**
- * Provides compatibility layer for logging.
- * @author Oleg Nenashev
- * @since TODO
- * @see jenkins.model.logging.Loggable
- * @see FileLogStorage
- * @see FileLogBrowser
- */
-@Restricted(Beta.class)
-public interface FileLogCompatLayer {
-
-    @Nonnull
-    default File getLogFileOrFail(Loggable loggable) throws IOException {
-        final File file = loggable.getLogFileCompatLocation();
-        if (file == null) {
-            throw new AbortException("File log compatibility layer is invoked for a loggable " +
-                    "object which returned null for getLogFileCompatLocation(): " + loggable);
-        }
-        return file;
-    }
-}
