--- core/src/main/java/jenkins/model/logging/impl/NoopLogBrowser.java
@@ -1,88 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2018 CloudBees, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package jenkins.model.logging.impl;
-
-import hudson.console.AnnotatedLargeText;
-import jenkins.model.logging.LogBrowser;
-import jenkins.model.logging.Loggable;
-import org.kohsuke.accmod.Restricted;
-import org.kohsuke.accmod.restrictions.Beta;
-
-import javax.annotation.CheckForNull;
-import javax.annotation.Nonnull;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.util.logging.Logger;
-
-/**
- * Default Log Browser implementation which does nothing.
- * @author Oleg Nenashev
- * @since TODO
- * @see NoopLoggingMethod
- */
-@Restricted(Beta.class)
-public class NoopLogBrowser extends LogBrowser {
-
-    private static final Logger LOGGER = Logger.getLogger(NoopLogBrowser.class.getName());
-
-    private transient File noopLogFile;
-
-    public NoopLogBrowser(Loggable loggable) {
-        super(loggable);
-    }
-
-    @Override
-    public AnnotatedLargeText overallLog() {
-        return new BrokenAnnotatedLargeText(
-                new UnsupportedOperationException("Browsing is not supported"),
-                getOwner().getCharset());
-    }
-
-    @Override
-    public AnnotatedLargeText stepLog(@CheckForNull String stepId, boolean b) {
-        return overallLog();
-    }
-
-    //TODO: It may be better to have a single file for all implementations, but Charsets may be different
-    @Nonnull
-    @Override
-    public File getLogFile() throws IOException {
-        if (noopLogFile == null) {
-            File f = File.createTempFile("deprecated", ".log", getOwner().getTmpDir());
-            f.deleteOnExit();
-            try (OutputStream os = new FileOutputStream(f)) {
-                overallLog().writeRawLogTo(0, os);
-            }
-            return f;
-        }
-        return noopLogFile;
-    }
-
-    @Override
-    public boolean deleteLog() {
-        return true;
-    }
-}
